+++ R/Consistency.R
@@ -41,12 +41,9 @@
 #' following \insertCite{Brazeau2019;textual}{TreeSearch}, matching their
 #' default treatment in [`TreeLength()`].
 #' 
-#' @template datasetParam
-#' @template treeParam
 #' @param nRelabel Integer specifying number of leaf relabellings to employ when
 #' calculating null tree length. If zero, the RHI will not be calculated.
-#' 
-#' @template compressParam
+#' @inheritParams CharacterLength
 #' 
 #' @return `Consistency()` returns a matrix with named columns specifying the 
 #' consistency index (`ci`),