+++ R/IWScore.R
@@ -1,4 +1,5 @@
-#' @template pointlessDots
+#' @param \dots unused; allows additional parameters specified within \dots to be 
+#'              received by the function without throwing an error.
 #' @rdname TreeLength
 #' @export
 IWScore <- function (tree, dataset, concavity = 10L, ...) {