+++ man-roxygen/datasetParam.r
@@ -1,3 +1,0 @@
-#' @param dataset A phylogenetic data matrix of \pkg{phangorn} class
-#' \code{phyDat}, whose names correspond to the labels of any accompanying tree.
-# Defined in TreeTools.  Please propagate any changes there.