+++ man-roxygen/gapParam.R
@@ -1,2 +1,0 @@
-#' @param gap An unambiguous abbreviation of `inapplicable`, `ambiguous` 
-#' (= `missing`), or `extra state`, specifying how gaps will be handled.