+++ man-roxygen/treeChild.R
@@ -1,3 +1,0 @@
-#' @param child Integer vector corresponding to the second column of the edge
-#' matrix of a tree of class \code{\link{phylo}}, i.e. `tree$edge[, 2]`.
-# Defined in TreeTools.  Please propagate any changes there.