+++ NEWS.md
@@ -1,7 +1,31 @@
-# TreeSearch 1.5.1.9000 (2024-05-31, development)
-
+# TreeSearch 1.6.1.9003 (development)
+- Support single-character matrices in `ClusteringConcordance()`
+- Fix `DoNothing(x)` to return `x` (not `NULL`)
 - `Consistency()` also returns the relative homoplasy index of Steell et al. 
   2023.
+
+# TreeSearch 1.6.1 (2025-06-10)
+- Handle invariant characters in `PolEscapa()`
+- Handle challenging root positions in `PlotCharacter()`
+- Fix character state colours in app legend
+- Tweak documentation
+
+# TreeSearch 1.6.0 (2025-04-09)
+
+## Improvements
+- `PlotCharacter()` performs ancestral state reconstruction on consensus trees
+  ([#179](https://github.com/ms609/TreeSearch/issues/179))
+- Improve support for constraints in `AdditionTree()`
+  ([#173](https://github.com/ms609/TreeSearch/issues/173))
+- Support for ordered (additive) characters via `TreeTools::Decompose()`
+- Fix SPR behaviour when move is close to root
+
+## App improvements
+- Buttons to download consensus trees in app
+- Fix display of state labels in app
+
+## Housekeeping
+- Require R 4.0 (to simplify maintenance)
 
 
 # TreeSearch 1.5.1 (2024-05-23)